Separate device.hpp into device-specific headers.
